task/cleanup-url-utils * replace getRestUrl with buildRestUrl in UrlBuilder
