+++ blueocean-core-js/src/js/index.js
@@ -104,4 +104,5 @@ export { TimeManager } from './utils/TimeManager';
 export { TimeHarmonizer, TimeHarmonizerUtil } from './components/TimeHarmonizer';
 import LiveStatusIndicator from './components/LiveStatusIndicator';
 export { LiveStatusIndicator };
-export { doubleUriEncode, fetchAllSuffix, ensureTrailingSlash, getRestUrl, buildUrl, relativeUrl } from './utils/UrlUtils';
+export { doubleUriEncode, fetchAllSuffix, ensureTrailingSlash, buildUrl, relativeUrl } from './utils/UrlUtils';
+// TODO: replace these with utils as namespace, add namespace to index test
