+++ blueocean-pipeline-editor/src/main/js/EditorPage.jsx
@@ -2,8 +2,7 @@ import React from 'react';
 import Extensions from '@jenkins-cd/js-extensions';
 import {
     Fetch,
-    getRestUrl,
-    buildPipelineUrl,
+    UrlBuilder,
     locationService,
     ContentPageHeader,
     pipelineService,
@@ -302,8 +301,8 @@ class PipelineLoader extends React.Component {
             const team = split[0];
             const repo = split.length > 1 ? split[1] : team;
             const { id: scmId, apiUrl } = this.state.scmSource;
-            // TODO: bitbucket isn't passing the pipeline in "orgname/reponame" format so this request 404's with bogus team name
-            let repositoryUrl = `${getRestUrl({ organization })}scm/${scmId}/organizations/${team}/repositories/${repo}/`;
+            const orgRestUrl = UrlBuilder.buildRestUrl(organization);
+            let repositoryUrl = `${orgRestUrl}scm/${scmId}/organizations/${team}/repositories/${repo}/`;
             if (apiUrl) {
                 repositoryUrl += `?apiUrl=${apiUrl}`;
             }
